+++ docs/misc/changelog.rst
@@ -6,6 +6,10 @@ Changelog
 Release 1.0rc1 (WIP)
 -------------------------------
 
+Breaking Changes:
+^^^^^^^^^^^^^^^^^
+- Upgraded to Stable-Baselines3 >= 1.0rc1
+
 Bug Fixes:
 ^^^^^^^^^^
 - Fixed a bug with ``QR-DQN`` predict method when using ``deterministic=False`` with image space